app/requirements.txt
djangorestframework
django-cors-headers
psycopg2-binary==2.9.1
google-api-python-client
oauth2client
